Update to discard changes on doc.close()
In our flow, the docx's created by our design team are marked as 'Read Only' so there's a flag set that will prompt to save changes even if you don't make any.  This prevented docx2pdf from succesfully working in batch mode.   

This fix was only prompted because (I think) something changed in Word.  it used to give me the option to not save the changes and keep moving